+++ src/app/app.routes.ts
@@ -1,7 +1,6 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './components/pages/home/home.component';
 import { DashboardComponent } from './components/pages/dashboard/dashboard.component';
-//import { LoginComponent } from './auth/login/login.component';
 import { LoginComponent } from './components/pages/login/login/login.component';
 import { AuthGuard } from './guards/auth.guard';
 import { TareaModalComponent } from './components/layout/tareas/tarea-modal/tarea-modal.component';
