+++ src/app/components/pages/login/login/login.component.spec.ts
@@ -11,7 +11,7 @@ describe('LoginComponent', () => {
       imports: [LoginComponent]
     })
     .compileComponents();
-    
+
     fixture = TestBed.createComponent(LoginComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
