+++ src/app/components/pages/login/login/login.component.ts
@@ -1,50 +1,58 @@
-import { FormsModule } from '@angular/forms';
-import { Component } from '@angular/core';
-import { AuthService } from '../../../../services/auth/auth.service';
+import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup, Validators} from '@angular/forms';
 import { Router } from '@angular/router';
+import { LoginService } from '../../../../services/auth/loginRequest';
+import { LoginRequest } from '../../../../services/auth/loginService';
 import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-login',
-  standalone: true,
-  imports: [FormsModule, CommonModule],
   templateUrl: './login.component.html',
-  styleUrl: './login.component.css',
-  providers: [AuthService],
+  styleUrls: ['./login.component.css'],
 })
-export class LoginComponent {
-  email: string = '';
-  password: string = '';
-  errorMessage: string = '';
+export class LoginComponent implements OnInit {
+  loginError:string="";
+  loginForm=this.formBuilder.group({
+    email:[' ',[Validators.required,Validators.email]],
+    password: ['',Validators.required],
+  })
+  constructor(private formBuilder:FormBuilder, private router:Router, private loginService: LoginService) { }
 
-  constructor(private authService: AuthService, private router: Router) {}
-
-  login() {
-    const usuarioData = {
-      email: this.email,
-      contraseña: this.password,
-    };
-    console.log(usuarioData);
-    this.authService.loginUsuario(usuarioData).subscribe(
-      (data) => {
-        console.log('Login successful', data);
-        localStorage.setItem('token', data.token);
-        this.router.navigate(['/dashboard']);
-      },
-      (error) => {
-        console.error('Login failed', error);
-        if (error.status === 401) {
-          this.errorMessage = 'Invalid email or password.';
-        } else if (error.status === 500) {
-          this.errorMessage =
-            'Your account is not activated. Please check your email for the activation link.';
-        } else if (error.status === 404) {
-          this.errorMessage = 'User not found.';
-        } else {
-          this.errorMessage =
-            'An unexpected error occurred. Please try again later.';
-        }
-      }
-    );
+  ngOnInit(): void {
   }
+
+  get email(){
+    return this.loginForm.controls.email;
+  }
+
+  get password()
+  {
+    return this.loginForm.controls.password;
+  }
+
+  login(){
+    if(this.loginForm.valid){
+      this.loginError="";
+      this.loginService.login(this.loginForm.value as LoginRequest).subscribe({
+        next: (userData) => {
+          console.log(userData);
+        },
+        error: (errorData) => {
+          console.error(errorData);
+          this.loginError=errorData;
+        },
+        complete: () => {
+          console.info("Login completo");
+          this.router.navigateByUrl('/inicio');
+          this.loginForm.reset();
+        }
+      })
+
+    }
+    else{
+      this.loginForm.markAllAsTouched();
+      alert("Error al ingresar los datos.");
+    }
+  }
+
 }
